implement cpu letterbox implementation, and store timer msgs as vector

+++ chapter7-deploy-yolo-detection/7.3-deploy-yolo-basic/src/cpp/main.cpp
@@ -16,7 +16,7 @@ int main(int argc, char const *argv[])
     params.img         = {640, 640, 3};
     params.task        = model::task_type::DETECTION;
     params.dev         = model::device::GPU;
-    params.prec        = model::precision::INT8;
+    params.prec        = model::precision::FP16;
 
     // 创建一个worker的实例, 在创建的时候就完成初始化
     auto worker   = thread::create_worker(onnxPath, level, params);
